refactor: 멀티플레이 지원을 위한 전체 패키지 구조 리팩토링
-  도메인 로직을 minesweeper-core 모듈로 이동
