+++ minesweeper-client/game/view/input.go
