+++ minesweeper-client/go.mod
@@ -3,8 +3,7 @@ module minesweeper-client
 go 1.23
 
 require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/stretchr/testify v1.11.1 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
+    minesweeper-core v0.0.0
 )
+
+replace minesweeper-core => ../minesweeper-core
