+++ minesweeper-core/cell/cell_snapshot.go
