+++ minesweeper-core/cell/cell_snapshot_status.go
