+++ minesweeper-core/cell/cell_state.go
