+++ minesweeper-core/cell/empty.go
