+++ minesweeper-core/cell/landmine.go
