+++ minesweeper-core/level/easy.go
