+++ minesweeper-core/position/cell_position.go
