+++ minesweeper-core/position/cell_position_test.go
