+++ minesweeper-core/position/cell_positions_test.go
@@ -1,7 +1,7 @@
 package position
 
 import (
-	"minesweeper-client/minesweeper/cell"
+	"minesweeper-core/cell"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
