+++ minesweeper-core/util/fatal.go
