+++ minesweeper-server/go.mod
@@ -1,3 +1,9 @@
 module minesweeper-server
 
 go 1.23
+
+require (
+    minesweeper-core v0.0.0
+)
+
+replace minesweeper-core => ../minesweeper-core
